+++ src/components/Features/CultureSelector.vue
@@ -1,130 +1,59 @@
 <template>
-  <div class="fr-input-group" :class="{ 'fr-input-group-error': requirePrecision }">
+  <fieldset class="fr-input-group" v-for="(culture) in uuidedCultures" :key="culture.id">
     <label class="fr-label">Type de culture</label>
-    <div v-if="requirePrecision && fromCodeCpf(modelValue)" class="fr-hint-text">
-      Culture «&nbsp;{{ fromCodeCpf(modelValue).libelle_code_cpf }}&nbsp;» à préciser
-    </div>
-    <div ref="autocompleteRef"></div>
-    <div v-if="requirePrecision" class="fr-hint-text fr-error-text">
-      La culture a besoin d'être précisée
-    </div>
-    <div v-else class="fr-hint-text">Saisissez le nom d'une culture</div>
-  </div>
+
+    <CultureTypeSelector :from-pac="culture.TYPE" :modelValue="culture.CPF" @update:modelValue="$CPF => updateCulture(culture.id, 'CPF', $CPF)" />
+
+    <button type="button" class="fr-btn fr-btn--sm fr-btn--tertiary-no-outline fr-icon-delete-line fr-btn--icon-left" :disabled="!canBeDeleted" @click="removeCulture(culture.id)">
+      Supprimer
+    </button>
+  </fieldset>
+
+  <button type="button" class="fr-btn fr-btn--tertiary-no-outline fr-icon-add-line fr-btn--icon-left" @click="appendEmptyCulture">
+    Ajouter une autre culture
+  </button>
 </template>
 
 <script setup>
-import { computed, Fragment, h, nextTick, onMounted, ref, render } from 'vue'
-import { autocomplete } from '@algolia/autocomplete-js'
-import '@algolia/autocomplete-theme-classic';
-import cpf from '@agencebio/rosetta-cultures/data/cpf.json'
-import { fromCodeCpf, fromCodePacAll } from "@agencebio/rosetta-cultures"
-
-const autocompleteRef = ref(null)
+import { computed } from 'vue'
+import CultureTypeSelector from './CultureTypeSelector.vue';
 
 const props = defineProps({
-  modelValue: {
-    type: String,
+  cultures: {
+    type: Array,
     required: true
-  },
-  fromPac: {
-    type: String,
-    required: false
   }
 })
-const emit = defineEmits(['update:modelValue'])
 
-const requirePrecision = computed(() => props.modelValue && !(fromCodeCpf(props.modelValue)?.is_selectable))
+const emit = defineEmits(['change'])
 
-const showMore = ref(false)
-
-const choices = computed(() => {
-  const selectableCpf = cpf.filter(({ is_selectable }) => is_selectable)
-
-  if (!requirePrecision.value || !props.fromPac || showMore.value) return selectableCpf
-
-  return fromCodePacAll(props.fromPac)
-      .filter(c => c.is_selectable)
+const uuidedCultures = computed(() => {
+  return props.cultures.map(culture => ({
+    ...culture,
+    id: culture.id ?? self.crypto.randomUUID()
+  }))
 })
 
-onMounted(() => {
-  const { setQuery } = autocomplete({
-    container: autocompleteRef.value,
-    placeholder: 'Saisissez le nom d’une culture',
-    openOnFocus: true,
-    getSources() {
-      return [
-        {
-          sourceId: 'cultures',
-          getItems ({ query }) {
-            const cultureChoices = choices.value
-                .filter(({ libelle_code_cpf: libelle }) => libelle.toLowerCase().includes((query || '').toLowerCase()))
-                .map(({ libelle_code_cpf: libelle, code_cpf: code }) => ({ code, libelle }))
-                .sort((a, b) => a.libelle.localeCompare(b.libelle))
+const canBeDeleted = computed(() => uuidedCultures.value.length > 1)
 
-            if (requirePrecision.value && !(showMore.value)) {
-              cultureChoices.push({
-                libelle: 'Voir toutes les cultures',
-                code: 'showMore'
-              })
-            }
+function appendEmptyCulture () {
+  const appendedCultures = [...uuidedCultures.value, { CPF: '', id: self.crypto.randomUUID() }]
 
-            return cultureChoices
-          },
-          templates: {
-            item ({ item, html }) {
-              if (item.code === 'showMore') {
-                return html`<span class="fr-link">Voir toutes les cultures</span>`
-              }
+  emit('change', appendedCultures)
+}
 
-              return item.libelle
-            }
+function removeCulture (cultureId) {
+  const updatedCultures = uuidedCultures.value.filter(({ id }) => id !== cultureId)
 
-          },
-          onSelect: function(event) {
-            if (event.item.code === 'showMore') {
-              showMore.value = true
-              event.setQuery('')
-              event.setIsOpen(true)
-              return nextTick(() => {
-                event.refresh()
-              })
-            }
+  emit('change', updatedCultures)
+}
 
-            event.setQuery(event.item.libelle);
-            emit('update:modelValue', event.item.code);
-          },
-        }
-      ]
-    },
-    renderer: { createElement: h, Fragment, render }
-  })
+function updateCulture (cultureId, field, value) {
+  const updatedCultures = uuidedCultures.value.map(culture => culture.id === cultureId ? ({
+    ...culture,
+    [field]: value
+  }) : culture)
 
-  setQuery?.(requirePrecision.value ? '' : fromCodeCpf(props.modelValue)?.libelle_code_cpf || '')
-})
+  emit('change', updatedCultures)
+}
 </script>
-
-<style>
-.aa-Panel {
-  z-index: 2000;
-}
-
-.aa-PanelLayout {
-  max-height: calc(100vh - 25rem);
-}
-
-.aa-Autocomplete {
-  margin-top: 0.5rem;
-}
-
-.aa-Item:hover {
-  background-color: #ececfe;
-}
-
-.aa-InputWrapperPrefix {
-  display: none;
-}
-
-.aa-InputWrapper {
-  padding-left: 0.75rem;
-}
-</style>
